+++ src/scad_klupe/base.clj
@@ -5,8 +5,10 @@
 ;;; module akin to iso.clj.
 
 (ns scad-klupe.base
-  (:require [scad-clj.model :as model]
-            [scad-tarmi.core :refer [sin cos τ long-hex-diagonal]]))
+  (:require [clojure.spec.alpha :as spec]
+            [scad-clj.model :as model]
+            [scad-tarmi.core :refer [sin cos τ long-hex-diagonal]]
+            [scad-klupe.schema.base :as base]))
 
 
 ;;;;;;;;;;;;;;;;;;;;;;;;
@@ -35,25 +37,20 @@
 ;; INTERFACE FUNCTIONS ;;
 ;;;;;;;;;;;;;;;;;;;;;;;;;
 
+;; The following is intended primarily as an interface for other,
+;; standards-specific parts of scad-klupe.
+
 (defn shank-section-lengths
   "Determine the lengths of the unthreaded and threaded parts of a bolt.
   These can be explicit in the parameters to this function, or else they
   are inferred from the total length and the length of the head."
   [{:keys [total unthreaded threaded head] :as parameters}]
-  {:pre [(number? head)
-         (not (neg? head))
-         (>= (or total head) head)]}
+  {:pre [(spec/valid? ::base/bolt-length-parameters parameters)]}
   (case (map some? [total unthreaded threaded])
-    [true  true  true ] (if (= total (+ unthreaded threaded head))
-                          [unthreaded threaded]
-                          (throw
-                            (ex-info "Contradictory bolt length parameters"
-                                     {:parameters parameters})))
+    [true  true  true ] [unthreaded threaded]
     [true  true  false] [unthreaded (- total unthreaded head)]
     [true  false false] [0 (- total head)]
-    [false false false] (throw
-                          (ex-info "Insufficient bolt length parameters"
-                                   {:parameters parameters}))
+    [false false false] [0 0]  ; Case is contradictory to spec.
     [false false true ] [0 threaded]
     [false true  true ] [unthreaded threaded]
     [true  false true ] [(- total threaded head) threaded]
@@ -61,7 +58,7 @@
 
 (defn bolt-length
   "Predict the overall length of a bolt, including the head.
-  Use shank-section-lengths partly as a sanity check on the inputs."
+  Use shank-section-lengths partly for its sanity check on the inputs."
   [{:keys [total head] :as parameters}]
   (let [[unthreaded threaded] (shank-section-lengths parameters)]
     (if total total (+ head unthreaded threaded))))