+++ CHANGELOG.md
@@ -8,5 +8,7 @@ This change log follows the conventions of [keepachangelog.com](http://keepachan
   - All interface functions take a map of options rather than a variary
     sequence.
   - No default bolt head type.
+- Added specs for parameters to the `bolt` function as well as bolt length
+  predictors.
 
 [Unreleased]: https://github.com/veikman/scad-klupe/compare/v0.1.0...HEAD